Refactor `socketio_utils` to `persona_utils` Move `list_personas` logic to a util method Support no current user in `list_personas` to retrieve all personas with no user filtering
